run-parallel-batches doesn't need a separate test file
